fixing unit tests.  Moved withLocale to separate helper file.
