+++ package.json
@@ -176,7 +176,7 @@
     "string-punctuation-tokenizer": "^2.2.0",
     "sudo-prompt": "6.2.1",
     "tc-electron-env": "0.9.0",
-    "tc-source-content-updater": "1.2.0-alpha.2",
+    "tc-source-content-updater": "1.2.0-alpha.4",
     "tc-strings": "0.1.7",
     "tc-tool": "4.1.0",
     "tc-ui-toolkit": "6.0.0",
