+++ src/js/containers/projectValidation/ProjectValidationContainer.js
@@ -19,7 +19,7 @@ import ProjectInformationCheck from '../../components/projectValidation/ProjectI
 import MergeConflictsCheck from '../../components/projectValidation/MergeConflictsCheck';
 import MissingVersesCheck from '../../components/projectValidation/MissingVersesCheck';
 import ProjectValidationNavigation from '../../components/projectValidation/ProjectValidationNavigation';
-import { withLocale } from '../Locale';
+import { withLocale } from '../../helpers/localeHelpers';
 
 class ProjectValidationContainer extends Component {
   render() {
