+++ src/js/reducers/resourcesReducer.js
@@ -1,3 +1,4 @@
+import { apiHelpers } from 'tc-source-content-updater';
 import consts from '../actions/ActionTypes';
 import * as Bible from '../common/BooksOfTheBible';
 import { addOwner } from '../helpers/ResourcesHelpers';
@@ -11,9 +12,9 @@ const initialState = {
 const resourcesReducer = (state = initialState, action) => {
   switch (action.type) {
   case consts.ADD_NEW_BIBLE_TO_RESOURCES:
-    // eslint-disable-next-line no-case-declarations
+  {
     const languageAndOwner = action.owner ? addOwner(action.languageId, action.owner) : action.languageId;
-    return {
+    const newState = {
       ...state,
       bibles: {
         ...state.bibles,
@@ -23,6 +24,15 @@ const resourcesReducer = (state = initialState, action) => {
         },
       },
     };
+
+    if (action.owner === apiHelpers.DOOR43_CATALOG) { // for backward compatibility, save as default
+      newState.bibles[action.languageId] = {
+        ...state.bibles[action.languageId],
+        [action.bibleId]: action.bibleData,
+      };
+    }
+    return newState;
+  }
   case consts.UPDATE_TARGET_VERSE:
     return {
       ...state,
